+++ yandex/resource_yandex_api_gateway_test.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"fmt"
 	"io/ioutil"
+	"os"
 	"testing"
 
 	"github.com/hashicorp/go-multierror"
@@ -153,6 +154,100 @@ func TestAccYandexAPIGateway_full(t *testing.T) {
 	})
 }
 
+func TestAccYandexAPIGateway_domainsUpdate(t *testing.T) {
+	t.Parallel()
+
+	testName := acctest.RandomWithPrefix("tf-api-gateway")
+	testDesc := acctest.RandomWithPrefix("tf-api-gateway-desc")
+	testCertificateId := getTestCertificateId(t)
+	testDomain1 := fmt.Sprintf("%s.tf-acc-tests.prod.apigwtest.ru", acctest.RandomWithPrefix("test1"))
+	testDomain2 := fmt.Sprintf("%s.tf-acc-tests.prod.apigwtest.ru", acctest.RandomWithPrefix("test2"))
+
+	// initial API Gateway creation
+	createConfig := fmt.Sprintf(`
+resource "yandex_api_gateway" "test-api-gateway" {
+  name        = "%s"
+  description = "%s"
+  custom_domains {
+    certificate_id = "%s"
+    fqdn = "%s"
+  }
+  spec = <<EOF
+%sEOF
+}`, testName, testDesc, testCertificateId, testDomain1, spec)
+	testCreateFunc := resource.TestStep{
+		Config: createConfig,
+		Check: resource.ComposeTestCheckFunc(
+			resource.TestCheckResourceAttr(apiGatewayResource, "custom_domains.0.certificate_id", testCertificateId),
+			resource.TestCheckResourceAttrSet(apiGatewayResource, "custom_domains.0.domain_id"),
+			resource.TestCheckResourceAttr(apiGatewayResource, "custom_domains.0.fqdn", testDomain1),
+			resource.TestCheckNoResourceAttr(apiGatewayResource, "custom_domains.1"),
+		),
+	}
+
+	// add domain
+	addDomainConfig := fmt.Sprintf(`
+resource "yandex_api_gateway" "test-api-gateway" {
+  name        = "%s"
+  description = "%s"
+  custom_domains {
+    certificate_id = "%s"
+    fqdn = "%s"
+  }
+  custom_domains {
+    certificate_id = "%s"
+    fqdn = "%s"
+  }
+  spec = <<EOF
+%sEOF
+}`, testName, testDesc, testCertificateId, testDomain1, testCertificateId, testDomain2, spec)
+	testAddDomainFunc := resource.TestStep{
+		Config: addDomainConfig,
+		Check: resource.ComposeTestCheckFunc(
+			resource.TestCheckResourceAttr(apiGatewayResource, "custom_domains.0.certificate_id", testCertificateId),
+			resource.TestCheckResourceAttrSet(apiGatewayResource, "custom_domains.0.domain_id"),
+			resource.TestCheckResourceAttr(apiGatewayResource, "custom_domains.0.fqdn", testDomain1),
+			resource.TestCheckResourceAttr(apiGatewayResource, "custom_domains.1.certificate_id", testCertificateId),
+			resource.TestCheckResourceAttrSet(apiGatewayResource, "custom_domains.1.domain_id"),
+			resource.TestCheckResourceAttr(apiGatewayResource, "custom_domains.1.fqdn", testDomain2),
+			resource.TestCheckNoResourceAttr(apiGatewayResource, "custom_domains.2"),
+		),
+	}
+
+	// remove domain
+	removeDomainConfig := fmt.Sprintf(`
+resource "yandex_api_gateway" "test-api-gateway" {
+  name        = "%s"
+  description = "%s"
+  custom_domains {
+    certificate_id = "%s"
+    fqdn = "%s"
+  }
+  spec = <<EOF
+%sEOF
+}`, testName, testDesc, testCertificateId, testDomain2, spec)
+	testRemoveDomainFunc := resource.TestStep{
+		Config: removeDomainConfig,
+		Check: resource.ComposeTestCheckFunc(
+			resource.TestCheckResourceAttr(apiGatewayResource, "custom_domains.0.certificate_id", testCertificateId),
+			resource.TestCheckResourceAttrSet(apiGatewayResource, "custom_domains.0.domain_id"),
+			resource.TestCheckResourceAttr(apiGatewayResource, "custom_domains.0.fqdn", testDomain2),
+			resource.TestCheckNoResourceAttr(apiGatewayResource, "custom_domains.1"),
+		),
+	}
+
+	resource.Test(t, resource.TestCase{
+		PreCheck:     func() { testAccPreCheck(t) },
+		Providers:    testAccProviders,
+		CheckDestroy: testYandexAPIGatewayDestroy,
+		Steps: []resource.TestStep{
+			testCreateFunc,
+			testAddDomainFunc,
+			testRemoveDomainFunc,
+		},
+	})
+}
+
 func basicYandexAPIGatewayTestStep(apiGatewayName, apiGatewayDesc, labelKey, labelValue string, spec string, apiGateway *apigateway.ApiGateway) resource.TestStep {
 	return resource.TestStep{
 		Config: testYandexAPIGatewayBasic(apiGatewayName, apiGatewayDesc, labelKey, labelValue, spec),
@@ -277,10 +372,12 @@ resource "yandex_api_gateway" "test-api-gateway" {
 }
 
 type testYandexAPIGatewayParameters struct {
-	name       string
-	desc       string
-	labelKey   string
-	labelValue string
+	name          string
+	desc          string
+	labelKey      string
+	labelValue    string
+	certificateId string
+	domain        string
 }
 
 func testYandexAPIGatewayFull(params testYandexAPIGatewayParameters) string {
@@ -301,4 +398,14 @@ resource "yandex_api_gateway" "test-api-gateway" {
 		params.labelKey,
 		params.labelValue,
 		spec)
+}
+
+func getTestCertificateId(t *testing.T) string {
+	certID := os.Getenv("APIGW_TEST_CERTIFICATE_ID")
+
+	if certID == "" {
+		t.Log("WARN: APIGW_TEST_CERTIFICATE_ID is not defined")
+	}
+
+	return certID
 }