+++ CHANGELOG.md
@@ -12,6 +12,7 @@ FEATURES:
 * clickhouse: added the feature to set maintenance_window when creating a cluster.
 * clickhouse: support cloud_storage with advanced settings
 * postgresql: add possibility to rename `yandex_mdb_postgresql_database`
+* api_gateway: add `custom_domains` attribute to `api_gateway` resource and data source
 
 ## 0.84.0 (December 16, 2022)
 BUG FIXES: